requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==1.0.0
google-generativeai==0.3.2
pydantic==1.10.12
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-multipart==0.0.6
aiofiles==23.2.1